pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tt-img-enc"
version = "1.0.0"
description = "ComfyUI custom node for automatic image/video conversion with steganography"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TT Tools", email = "support@tttools.com"}
]
keywords = ["comfyui", "node", "image", "video", "conversion", "steganography"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.7"
dependencies = [
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "numpy>=1.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
]

[project.urls]
Homepage = "https://github.com/tttools/tt-img-enc"
Repository = "https://github.com/tttools/tt-img-enc.git"
Issues = "https://github.com/tttools/tt-img-enc/issues"

[tool.setuptools]
packages = ["tt_img_enc"]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=tt_img_enc",
    "--cov-report=term-missing",
    "--cov-report=html",
]
